Add function for listing test images
`list_images` is similar to list_csvs in vic2d.py in that it
specifically lists images with names following the capture format
(cam#_#####_###.###.tiff).  This is useful because there are sometimes
extra images in the images folder that should not be included in an
image sequence for e.g.  Vic-2D.  For example, pictures of the ruptured
specimen after the test is over.