ssview: Set colormap scale using quantiles, fix +/- inversion
The quantiles are calculated for each strain field individually,
then the maximum of absolute value is chosen for each of exx,
eyy, and ezz and used for the entire dataset.

The positive and negative strains were inverted when makeARGB was
applied; this is because pyqtgraph expects RGBA images.  Note that
makeARGB actually stores data in BGRA order (this is a Qt quirk).