Fix offset time error in ssview
Previously, ssview assumed that the first image in the provided
image list was the frame at mechanical t = 0.  This is not the case
if motion starts after a delay.  Early ramp to failure tests had
a 15 s hold followed by a ramp; this bug affected the display
of those tests.

Now, ssview looks up the ref_time image in the image index, then looks
for ref_time.csv (which records the timestamp of that frame from the
perspective of the mechanical tester) and uses this information to
compute the correct offset.

Note that this fix duplicates code in key_stress_pts(), in the same
file.