Add function to tabulate images and csv files
The new function is mechana.images.tabulate_images.  In additional to
collecting the test images and vic-2d file paths in a single table, it
is also meant to take care of matching the image timestamps to the
mechanical data.  Accordingly, mechana.analysis.MechanicalTest (which
ssview uses) was modified to use tabulate_images.