Instron CSV files enclose units with parentheses
mechana.read.instron_data is deprecated anyway, but some old code relies
on it, so it may as well work.  I previously updated mechana to output
tables that used column names with units in square brackets.  In the
process, I inadvertently updated the deprecated code that reads
Bluehill (Instron) csv files to expect the same, but of course Bluehill
uses parentheses to enclose units.